SITES-13120 : move interface to new location.
