+++ content/src/content/jcr_root/apps/core/wcm/components/image/v3/image/nextgendmthumbnail/nextgendmthumbnail.html
@@ -13,6 +13,6 @@
   ~ See the License for the specific language governing permissions and
   ~ limitations under the License.
   ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~*/-->
-<sly data-sly-use.thumbnail="com.adobe.cq.wcm.core.components.models.nextgendm.NextGenDMThumbnail" data-sly-use.inheritedField="com.adobe.cq.wcm.core.components.commons.editor.dialog.inherited.InheritedField">
+<sly data-sly-use.thumbnail="com.adobe.cq.wcm.core.components.commons.editor.nextgendm.NextGenDMThumbnail" data-sly-use.inheritedField="com.adobe.cq.wcm.core.components.commons.editor.dialog.inherited.InheritedField">
      <img src="${thumbnail.src}"  class="${inheritedField.attrClass}" alt="${thumbnail.alt}"/>
 </sly>
